chore: moves utils around
